Move filters to inference; distribute handlers
With corresponding doc build changes
