--- dynestyx/inference/continuous_time_filters.py
@@ -1,45 +1,0 @@
-"""Continuous-time filter dispatcher: delegates to cd-dynamax (and future backends)."""
-
-import jax
-
-from dynestyx.inference.filter_configs import BaseFilterConfig
-from dynestyx.inference.integrations.cd_dynamax.continuous import run_continuous_filter
-from dynestyx.models import DynamicalModel
-
-
-def _filter_continuous_time(
-    name: str,
-    dynamics: DynamicalModel,
-    filter_config: BaseFilterConfig,
-    key: jax.Array | None = None,
-    *,
-    obs_times: jax.Array,
-    obs_values: jax.Array,
-    ctrl_times=None,
-    ctrl_values=None,
-    **kwargs,
-) -> None:
-    """Continuous-time marginal likelihood via CD-Dynamax.
-
-    Supports: EnKF, DPF, EKF, UKF (inferred from config type).
-
-    Args:
-        name: Name of the factor.
-        dynamics: Dynamical model to filter.
-        filter_config: Configuration for the filter.
-        obs_times: Observation times.
-        obs_values: Observed values.
-        ctrl_times: Control times (optional).
-        ctrl_values: Control values (optional).
-    """
-    run_continuous_filter(
-        name,
-        dynamics,
-        filter_config,  # type: ignore[arg-type]
-        key=key,
-        obs_times=obs_times,
-        obs_values=obs_values,
-        ctrl_times=ctrl_times,
-        ctrl_values=ctrl_values,
-        **kwargs,
-    )
