--- dynestyx/inference/discrete_time_filters.py
@@ -1,66 +1,0 @@
-"""Discrete-time filter dispatcher: delegates to cuthbert or cd-dynamax backends."""
-
-import jax
-
-from dynestyx.inference.filter_configs import BaseFilterConfig
-from dynestyx.inference.integrations.cd_dynamax.discrete import (
-    run_discrete_filter as run_cd_dynamax_discrete,
-)
-from dynestyx.inference.integrations.cuthbert.discrete import (
-    run_discrete_filter as run_cuthbert_discrete,
-)
-from dynestyx.models import DynamicalModel
-
-
-def _filter_discrete_time(
-    name: str,
-    dynamics: DynamicalModel,
-    filter_config: BaseFilterConfig,
-    key: jax.Array | None = None,
-    *,
-    obs_times: jax.Array,
-    obs_values: jax.Array,
-    ctrl_times=None,
-    ctrl_values=None,
-    **kwargs,
-) -> None:
-    """Discrete-time marginal likelihood via cuthbert or cd-dynamax.
-
-    Filter type inferred from config class: KFConfig, EKFConfig, UKFConfig (cd-dynamax)
-    or EKFConfig (cuthbert), PFConfig (cuthbert).
-
-    Args:
-        name: Name of the factor.
-        dynamics: Dynamical model to filter.
-        filter_config: Configuration for the filter.
-        obs_times: Observation times.
-        obs_values: Observed values.
-        ctrl_times: Control times (optional).
-        ctrl_values: Control values (optional).
-    """
-
-    if filter_config.filter_source == "cd_dynamax":
-        run_cd_dynamax_discrete(
-            name,
-            dynamics,
-            filter_config,
-            obs_times=obs_times,
-            obs_values=obs_values,
-            ctrl_times=ctrl_times,
-            ctrl_values=ctrl_values,
-            **kwargs,
-        )
-    elif filter_config.filter_source == "cuthbert":
-        run_cuthbert_discrete(
-            name,
-            dynamics,
-            filter_config,
-            key=key,
-            obs_times=obs_times,
-            obs_values=obs_values,
-            ctrl_times=ctrl_times,
-            ctrl_values=ctrl_values,
-            **kwargs,
-        )
-    else:
-        raise ValueError(f"Unknown filter source: {filter_config.filter_source}")
