+++ dynestyx/inference/filters.py
@@ -3,12 +3,10 @@ import dataclasses
 import jax
 import numpyro
 from cd_dynamax import ContDiscreteNonlinearGaussianSSM, ContDiscreteNonlinearSSM
+from effectful.ops.semantics import fwd
+from effectful.ops.syntax import ObjectInterpretation, implements
 
-from dynestyx.handlers import BaseCDDynamaxLogFactorAdder
-from dynestyx.inference.continuous_time_filters import (
-    _filter_continuous_time,
-)
-from dynestyx.inference.discrete_time_filters import _filter_discrete_time
+from dynestyx.handlers import HandlesSelf, sample
 from dynestyx.inference.filter_configs import (
     BaseFilterConfig,
     ContinuousTimeConfigs,
@@ -28,9 +26,68 @@ from dynestyx.inference.filter_configs import (
     UKFConfig,
 )
 from dynestyx.inference.hmm_filters import _filter_hmm
+from dynestyx.inference.integrations.cd_dynamax.continuous import run_continuous_filter
+from dynestyx.inference.integrations.cd_dynamax.discrete import (
+    run_discrete_filter as run_cd_dynamax_discrete,
+)
+from dynestyx.inference.integrations.cuthbert.discrete import (
+    run_discrete_filter as run_cuthbert_discrete,
+)
 from dynestyx.models import DynamicalModel
+from dynestyx.types import FunctionOfTime
 
 type SSMType = ContDiscreteNonlinearGaussianSSM | ContDiscreteNonlinearSSM
+
+
+class BaseLogFactorAdder(ObjectInterpretation, HandlesSelf):
+    """Base for filter handlers."""
+
+    @implements(sample)
+    def _sample_ds(
+        self,
+        name: str,
+        dynamics: DynamicalModel,
+        *,
+        obs_times=None,
+        obs_values=None,
+        ctrl_times=None,
+        ctrl_values=None,
+        **kwargs,
+    ) -> FunctionOfTime:
+        self.add_log_factors(
+            name,
+            dynamics,
+            obs_times=obs_times,
+            obs_values=obs_values,
+            ctrl_times=ctrl_times,
+            ctrl_values=ctrl_values,
+            **kwargs,
+        )
+        # Forward unchanged so downstream handlers (or default implementation)
+        # can still see this op if needed.
+        return fwd(
+            name,
+            dynamics,
+            obs_times=obs_times,
+            obs_values=obs_values,
+            ctrl_times=ctrl_times,
+            ctrl_values=ctrl_values,
+            **kwargs,
+        )
+
+    def add_log_factors(
+        self,
+        name: str,
+        dynamics: DynamicalModel,
+        *,
+        obs_times=None,
+        obs_values=None,
+        ctrl_times=None,
+        ctrl_values=None,
+        **kwargs,
+    ):
+        # Inheritors should implement this method.
+        raise NotImplementedError()
 
 
 def _default_filter_config(dynamics: DynamicalModel):
@@ -43,7 +100,7 @@ def _default_filter_config(dynamics: DynamicalModel):
 
 
 @dataclasses.dataclass
-class Filter(BaseCDDynamaxLogFactorAdder):
+class Filter(BaseLogFactorAdder):
     """
     Object for filtering a dynamical model, and adding the resulting marginal log likelihood as a numpyro factor.
 
@@ -141,6 +198,98 @@ class Filter(BaseCDDynamaxLogFactorAdder):
                     f"Invalid filter config: {type(config).__name__}. "
                     f"Valid config types: {valid}"
                 )
+
+
+def _filter_discrete_time(
+    name: str,
+    dynamics: DynamicalModel,
+    filter_config: BaseFilterConfig,
+    key: jax.Array | None = None,
+    *,
+    obs_times: jax.Array,
+    obs_values: jax.Array,
+    ctrl_times=None,
+    ctrl_values=None,
+    **kwargs,
+) -> None:
+    """Discrete-time marginal likelihood via cuthbert or cd-dynamax.
+
+    Filter type inferred from config class: KFConfig, EKFConfig, UKFConfig (cd-dynamax)
+    or EKFConfig (cuthbert), PFConfig (cuthbert).
+
+    Args:
+        name: Name of the factor.
+        dynamics: Dynamical model to filter.
+        filter_config: Configuration for the filter.
+        obs_times: Observation times.
+        obs_values: Observed values.
+        ctrl_times: Control times (optional).
+        ctrl_values: Control values (optional).
+    """
+
+    if filter_config.filter_source == "cd_dynamax":
+        run_cd_dynamax_discrete(
+            name,
+            dynamics,
+            filter_config,
+            obs_times=obs_times,
+            obs_values=obs_values,
+            ctrl_times=ctrl_times,
+            ctrl_values=ctrl_values,
+            **kwargs,
+        )
+    elif filter_config.filter_source == "cuthbert":
+        run_cuthbert_discrete(
+            name,
+            dynamics,
+            filter_config,
+            key=key,
+            obs_times=obs_times,
+            obs_values=obs_values,
+            ctrl_times=ctrl_times,
+            ctrl_values=ctrl_values,
+            **kwargs,
+        )
+    else:
+        raise ValueError(f"Unknown filter source: {filter_config.filter_source}")
+
+
+def _filter_continuous_time(
+    name: str,
+    dynamics: DynamicalModel,
+    filter_config: BaseFilterConfig,
+    key: jax.Array | None = None,
+    *,
+    obs_times: jax.Array,
+    obs_values: jax.Array,
+    ctrl_times=None,
+    ctrl_values=None,
+    **kwargs,
+) -> None:
+    """Continuous-time marginal likelihood via CD-Dynamax.
+
+    Supports: EnKF, DPF, EKF, UKF (inferred from config type).
+
+    Args:
+        name: Name of the factor.
+        dynamics: Dynamical model to filter.
+        filter_config: Configuration for the filter.
+        obs_times: Observation times.
+        obs_values: Observed values.
+        ctrl_times: Control times (optional).
+        ctrl_values: Control values (optional).
+    """
+    run_continuous_filter(
+        name,
+        dynamics,
+        filter_config,  # type: ignore[arg-type]
+        key=key,
+        obs_times=obs_times,
+        obs_values=obs_values,
+        ctrl_times=ctrl_times,
+        ctrl_values=ctrl_values,
+        **kwargs,
+    )
 
 
 __all__ = [
